+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
+- `Exception::$errorInfo` is now private
+- `attributes()` is now `getAttributes()`
 
 ## [0.7.0] - 2016-02-29
 ### Added